+++ packages/core/src/services/level-height.ts
@@ -1,6 +1,7 @@
-import { pointInPolygon } from '../hooks/spatial-grid/spatial-grid-manager'
-import type { CeilingNode, LevelNode, WallNode } from '../schema'
+import type { CeilingNode, LevelNode, SlabNode, WallNode } from '../schema'
 import type { AnyNode, AnyNodeId } from '../schema/types'
+import { computeWallSlabSupport, pointInPolygon } from '../systems/slab/slab-support'
+import { DEFAULT_WALL_HEIGHT } from '../systems/wall/wall-footprint'
 
 export const DEFAULT_LEVEL_HEIGHT = 2.5
 
@@ -35,6 +36,45 @@ export function getLevelHeight(
       let baseY = resolveWallBaseY?.(childId as AnyNodeId) ?? 0
       if (baseY < 0) baseY = 0
       const top = baseY + ((child as WallNode).height ?? DEFAULT_LEVEL_HEIGHT)
+      if (top > maxTop) maxTop = top
+    }
+  }
+
+  return maxTop > 0 ? maxTop : DEFAULT_LEVEL_HEIGHT
+}
+
+export function deriveLegacyLevelHeight(
+  levelId: string,
+  nodes: Record<AnyNodeId, AnyNode>,
+): number {
+  const level = nodes[levelId as LevelNode['id']] as LevelNode | undefined
+  if (!level) return DEFAULT_LEVEL_HEIGHT
+
+  const levelChildren = level.children
+    .map((childId) => nodes[childId as keyof typeof nodes])
+    .filter((child): child is AnyNode => child !== undefined)
+  const slabs = levelChildren.filter((child): child is SlabNode => child.type === 'slab')
+  const walls = levelChildren.filter((child): child is WallNode => child.type === 'wall')
+
+  let maxTop = 0
+
+  for (const child of levelChildren) {
+    if (child.type === 'ceiling') {
+      const height = (child as CeilingNode).height ?? DEFAULT_LEVEL_HEIGHT
+      if (height > maxTop) maxTop = height
+    } else if (child.type === 'wall') {
+      const wall = child as WallNode
+      const electedElevation = computeWallSlabSupport(
+        {
+          start: wall.start,
+          end: wall.end,
+          curveOffset: wall.curveOffset,
+          thickness: wall.thickness,
+        },
+        slabs,
+        walls,
+      ).elevation
+      const top = Math.max(0, electedElevation) + (wall.height ?? DEFAULT_WALL_HEIGHT)
       if (top > maxTop) maxTop = top
     }
   }
